RAM usage much improved
Adds a loop that communicates with QEMU via a socket to command the
balloon driver, which was previously doing nothing. Now the VM tries to
maintain a set level of free RAM for Linux by limiting amount of RAM
available to Windows, evaluating the situation once a second.

This adds a nice speed boost, because beforehand the way this was
working when literally anything else was running on the desktop, was due
to ZRAM. So this also fixes the "Killed" message some non-ZRAM-users
were getting.
New dependency added: socat to communicate with qemu
And new config file option: free_ram_goal
Now allocate 6GB on 8GB, and 14GB on 16GB devices, instead of 7GB and
15GB as before. Helps prevent oomkiller incidents.